requirements.txt
gradio 
datasets 
pillow 
requests
numpy